requirements.txt
streamlit
pandas
sentence-transformers
torch
pdfplumber
python-docx
numpy